// shared/schema.d.ts
export declare const insertUserSchema: z.ZodObject<Omit<{
    id: z.ZodOptional<z.ZodNumber>;
    username: z.ZodString;
    password: z.ZodString;
    name: z.ZodString;
    email: z.ZodString;
    createdAt: z.ZodOptional<z.ZodDate>;
    updatedAt: z.ZodOptional<z.ZodDate>;
}, "id" | "createdAt" | "updatedAt">, "strip", z.ZodTypeAny, {
    username: string;
    email: string;
    password: string;
    name: string;
}, {
    username: string;
    email: string;
    password: string;
    name: string;
}>;